Fix an error that caused deductions with negative numbers to be impossible
This was done by adding an additional if clause to getOperationElements().
When a new '-' element is entered, the function first checks if such an element
is already included in the first operator and, if it is, it adds it as an
operator, progressing the if clause tree to the last one.