Fix decimal error on 2nd operand
The error consisted in the user being unable to add back a decimal after
deciding to remove it with backspace.

The error was fixed by adding an additional check for '.' to the 2nd operand
in the operationElements array and by only parasin them to floating numbers
in the getResult() function instead of the getOperationElements() function.